refactor(tracker-client): split tracker_checker integration tests by concern
