+++ console/tracker-client/tests/tracker_checker.rs
@@ -47,248 +47,8 @@ fn resolve_tracker_checker_binary() -> std::path::PathBuf {
     );
 }
 
-mod invalid_configuration_from_env_var {
-    use super::tracker_checker_bin;
+#[path = "tracker_checker/configuration.rs"]
+mod configuration;
 
-    #[test]
-    fn it_should_exit_with_code_2_on_invalid_json() {
-        let output = tracker_checker_bin()
-            .env("TORRUST_CHECKER_CONFIG", r#"{"invalid json":"#)
-            .output()
-            .expect("Failed to run tracker_checker");
-
-        assert_eq!(output.status.code(), Some(2), "Expected exit code 2 for invalid config");
-    }
-
-    #[test]
-    fn it_should_write_json_error_to_stderr_on_invalid_json() {
-        let output = tracker_checker_bin()
-            .env("TORRUST_CHECKER_CONFIG", r#"{"invalid json":"#)
-            .output()
-            .expect("Failed to run tracker_checker");
-
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        assert!(
-            stderr.contains(r#""kind":"invalid_configuration""#),
-            "Expected JSON error envelope on stderr, got: {stderr}"
-        );
-        assert!(
-            stderr.contains(r#""source":"TORRUST_CHECKER_CONFIG""#),
-            "Expected source field to identify env var, got: {stderr}"
-        );
-    }
-
-    #[test]
-    fn it_should_include_parse_detail_in_stderr_error_message_on_trailing_comma() {
-        let config = r#"{
-            "udp_trackers": [],
-            "http_trackers": [
-                "http://127.0.0.1:7070",
-            ],
-            "health_checks": []
-        }"#;
-
-        let output = tracker_checker_bin()
-            .env("TORRUST_CHECKER_CONFIG", config)
-            .output()
-            .expect("Failed to run tracker_checker");
-
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        assert_eq!(output.status.code(), Some(2), "Expected exit code 2 for invalid config");
-        assert!(
-            stderr.contains("trailing comma"),
-            "Expected 'trailing comma' detail in stderr, got: {stderr}"
-        );
-    }
-
-    #[test]
-    fn it_should_produce_no_output_on_stdout_on_config_error() {
-        let output = tracker_checker_bin()
-            .env("TORRUST_CHECKER_CONFIG", r#"{"invalid json":"#)
-            .output()
-            .expect("Failed to run tracker_checker");
-
-        // Per the I/O contract, stdout is for successful results only
-        let stdout = String::from_utf8_lossy(&output.stdout);
-        assert!(stdout.is_empty(), "Expected no stdout on config error, got: {stdout}");
-    }
-}
-
-mod invalid_configuration_from_file {
-    use std::io::Write;
-
-    use super::tracker_checker_bin;
-
-    #[test]
-    fn it_should_exit_with_code_2_on_invalid_json_in_file() {
-        let mut tmp = tempfile::NamedTempFile::new().expect("Failed to create temp file");
-        write!(tmp, r#"{{"invalid json":"#).unwrap();
-
-        let output = tracker_checker_bin()
-            .env("TORRUST_CHECKER_CONFIG_PATH", tmp.path())
-            .output()
-            .expect("Failed to run tracker_checker");
-
-        assert_eq!(output.status.code(), Some(2), "Expected exit code 2 for invalid config file");
-    }
-
-    #[test]
-    fn it_should_include_file_path_in_stderr_source_field() {
-        let mut tmp = tempfile::NamedTempFile::new().expect("Failed to create temp file");
-        write!(tmp, r#"{{"invalid json":"#).unwrap();
-        let path = tmp.path().to_string_lossy().to_string();
-
-        let output = tracker_checker_bin()
-            .env("TORRUST_CHECKER_CONFIG_PATH", &path)
-            .output()
-            .expect("Failed to run tracker_checker");
-
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        assert!(
-            stderr.contains(&path),
-            "Expected file path in stderr source field, got: {stderr}"
-        );
-    }
-
-    #[test]
-    fn it_should_exit_with_code_2_when_config_file_does_not_exist() {
-        let output = tracker_checker_bin()
-            .env("TORRUST_CHECKER_CONFIG_PATH", "/nonexistent/path/config.json")
-            .output()
-            .expect("Failed to run tracker_checker");
-
-        assert_eq!(output.status.code(), Some(2), "Expected exit code 2 for missing config file");
-    }
-}
-
-mod no_configuration_provided {
-    use super::tracker_checker_bin;
-
-    #[test]
-    fn it_should_exit_with_code_2_when_no_config_is_provided() {
-        let output = tracker_checker_bin()
-            // Ensure neither env var is set
-            .env_remove("TORRUST_CHECKER_CONFIG")
-            .env_remove("TORRUST_CHECKER_CONFIG_PATH")
-            .output()
-            .expect("Failed to run tracker_checker");
-
-        assert_eq!(output.status.code(), Some(2), "Expected exit code 2 when no config provided");
-    }
-
-    #[test]
-    fn it_should_write_json_error_to_stderr_when_no_config_is_provided() {
-        let output = tracker_checker_bin()
-            .env_remove("TORRUST_CHECKER_CONFIG")
-            .env_remove("TORRUST_CHECKER_CONFIG_PATH")
-            .output()
-            .expect("Failed to run tracker_checker");
-
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        assert!(
-            stderr.contains(r#""kind":"invalid_configuration""#),
-            "Expected JSON error envelope on stderr, got: {stderr}"
-        );
-    }
-}
-
-/// Tests for the `monitor udp` subcommand.
-///
-/// # Timeout-only test environment
-///
-/// The helper [`spawn_udp_sink`] binds a UDP socket that silently discards every incoming
-/// packet and never sends any response. This means every probe issued by the monitor will
-/// time out. The tests in this module therefore exercise:
-///
-/// - JSON shape of probe events on stderr (`"status":"timeout"`)
-/// - JSON shape of the final summary on stdout (null latency fields, `timeout_percent` > 0)
-/// - Exit code 0 for a completed-but-all-timeout run
-///
-/// They do **not** exercise the success path (a probe receiving a valid `AnnounceResponse`,
-/// non-null `elapsed_ms`, populated min/max/average latency stats). A success-path
-/// integration test requires a proper mock UDP tracker that speaks the `BitTorrent` UDP
-/// protocol. See refactor plan item 14 in
-/// `docs/refactor-plans/open/1178-monitor-udp-post-implementation-improvements.md`.
-mod monitor_udp {
-    use std::net::{SocketAddr, UdpSocket};
-    use std::sync::mpsc;
-    use std::thread;
-    use std::time::Duration;
-
-    use serde_json::Value;
-
-    use super::tracker_checker_bin;
-
-    fn spawn_udp_sink() -> (SocketAddr, mpsc::Sender<()>, thread::JoinHandle<()>) {
-        let socket = UdpSocket::bind("127.0.0.1:0").expect("Failed to bind UDP sink socket");
-        socket
-            .set_read_timeout(Some(Duration::from_millis(100)))
-            .expect("Failed to configure UDP sink read timeout");
-        let addr = socket.local_addr().expect("Failed to get UDP sink local address");
-
-        let (tx, rx) = mpsc::channel::<()>();
-        let join_handle = thread::spawn(move || {
-            let mut buffer = [0_u8; 2048];
-
-            loop {
-                if rx.try_recv().is_ok() {
-                    break;
-                }
-
-                drop(socket.recv_from(&mut buffer));
-            }
-        });
-
-        (addr, tx, join_handle)
-    }
-
-    #[test]
-    fn it_should_emit_monitor_probe_events_to_stderr_and_summary_to_stdout() {
-        let (addr, stop_tx, join_handle) = spawn_udp_sink();
-
-        let output = tracker_checker_bin()
-            .arg("monitor")
-            .arg("udp")
-            .arg("--url")
-            .arg(format!("udp://{addr}"))
-            .arg("--interval")
-            .arg("1")
-            .arg("--timeout")
-            .arg("1")
-            .arg("--duration")
-            .arg("1")
-            .output()
-            .expect("Failed to run tracker_checker monitor udp");
-
-        let _ = stop_tx.send(());
-        drop(join_handle.join());
-
-        assert_eq!(
-            output.status.code(),
-            Some(0),
-            "Expected exit code 0 for successful monitor execution"
-        );
-
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        assert!(
-            stderr.contains("\"event\":\"probe\""),
-            "Expected probe NDJSON events on stderr, got: {stderr}"
-        );
-
-        let stdout = String::from_utf8_lossy(&output.stdout);
-        let parsed: Value = serde_json::from_str(&stdout).expect("Expected valid JSON monitor summary on stdout");
-
-        assert!(
-            parsed["udp_trackers"].is_array(),
-            "Expected udp_trackers array in stdout JSON"
-        );
-        assert_eq!(parsed["udp_trackers"][0]["url"], format!("udp://{addr}"));
-        assert!(
-            parsed["udp_trackers"][0]["status"]["stats"]["total"]
-                .as_u64()
-                .expect("Expected stats.total to be u64")
-                >= 1,
-            "Expected at least one probe"
-        );
-    }
-}
+#[path = "tracker_checker/monitor.rs"]
+mod monitor;
